requirements.txt
langchain==0.0.244
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.25.0
openai==0.27.8
faiss-cpu==1.7.4
altair<5
tiktoken
pycryptodome